Commit to a fixed PutBlob->PutManifest->PutSignatures ordering in ImageDestination
This is really the only ordering which makes sense, so document it and
remove various FIXMEs about the interdependencies.

Then reorder the implementations to use the same method order, for
readability.

To verify that this commit only changes comments and ordering, and
a single error message woring, check out this commit and use
> for i in $(git diff --name-only HEAD^ HEAD); do (echo $i; diff -U9 <(git show HEAD^:$i|LC_ALL=C sort) <(git show HEAD:$i|LC_ALL=C sort)) | less; done

Signed-off-by: Miloslav Trmač <mitr@redhat.com>