Add ImageDestination.SupportsSignatures
This allows copy.go to bail early, without transferring any large blobs.

(This is an attribute of an ImageDestination and not an ImageTransport
because supporting signatures may depend on the individual destination,
e.g. depending on server version or whether a sigstore lookaside has
been configured for that destination.)

Also reorders the ImageDestination methods in some cases, to separate
basic management / capability query / upload groups.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>