Split image.UnparsedImage from image.genericImage
Now that types.UnparsedImage is split from types.Image, also split the
implementations.  In particular, copy.Image uses an UnparsedImage for
signature verification.

This structural separation allows us to remove the “It is essential for
signature verification”… comments all over the place in favor of a
single one in the choke point where an UnparsedImage turns into a full
genericImage.

Also, split the manifest type guessing (which involves parsing) so that
it does not happen in UnparsedImage.  This needs ugly fields like
trueManifestMIMETypeSet, which will go away momentarily.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>