oci: remove MIME type autodetection
OCI JSON blobs no longer contain a mediaType (since 1.0.0-rc4), which
breaks the auto-detection code for MIME types. However, it doesn't make
sense for us to be "auto-detecting" the MIME type of a manifest when the
OCI transport *already knows the mediaType* from the reference. So
instead just return the already-computed mediaType from the descriptor.

Unfortunately, this doesn't work for GetTargetManifest, in which case we
just default to the only valid value at the moment. In future we should
extend this logic to read `/index.json` and figure out what descriptor
referenced the manifest (though any such logic would be questionable
without running a full index on all blobs in the source image).

Signed-off-by: Aleksa Sarai <asarai@suse.de>